git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cmav@20322 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ scripts/benchmark/vcloud.py
@@ -23,41 +23,58 @@ CPAchecker web page:
 """
 
 # prepare for Python 3
-from __future__ import absolute_import, print_function, unicode_literals
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 import sys
 sys.dont_write_bytecode = True # prevent creation of .pyc files
 
+import json
 import logging
 import os
 import subprocess
 import time
 
-from .benchmarkDataStructures import MEMLIMIT, TIMELIMIT, CORELIMIT
-from . import filewriter as filewriter
-from . import util as Util
+from benchexec.model import MEMLIMIT, TIMELIMIT, CORELIMIT
+import benchexec.util as util
 
 
 DEFAULT_CLOUD_TIMELIMIT = 300 # s
 DEFAULT_CLOUD_MEMLIMIT = None
 
-DEFAULT_CLOUD_MEMORY_REQUIREMENT = 7000 # MB
+DEFAULT_CLOUD_MEMORY_REQUIREMENT = 7000000000 # 7 GB
 DEFAULT_CLOUD_CPUCORE_REQUIREMENT = 2 # one core with hyperthreading
 DEFAULT_CLOUD_CPUMODEL_REQUIREMENT = "" # empty string matches every model
 
 STOPPED_BY_INTERRUPT = False
 
+_ROOT_DIR=os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir, os.pardir))
 
-def executeBenchmarkInCloud(benchmark, outputHandler, justReprocessResults):
-    if not justReprocessResults:
+_justReprocessResults = False
+
+def init(config, benchmark):
+    global _justReprocessResults
+    _justReprocessResults = config.reprocessResults
+    benchmark.executable = benchmark.tool.executable()
+    benchmark.tool_version = benchmark.tool.version(benchmark.executable)
+
+def get_system_info():
+    return None
+
+def execute_benchmark(benchmark, output_handler):
+    if not _justReprocessResults:
         # build input for cloud
         (cloudInput, numberOfRuns) = getCloudInput(benchmark)
-        cloudInputFile = os.path.join(benchmark.logFolder, 'cloudInput.txt')
-        filewriter.writeFile(cloudInput, cloudInputFile)
-        outputHandler.allCreatedFiles.append(cloudInputFile)
+        if benchmark.config.debug:
+            cloudInputFile = os.path.join(benchmark.log_folder, 'cloudInput.txt')
+            util.write_file(cloudInput, cloudInputFile)
+            output_handler.all_created_files.add(cloudInputFile)
+        meta_information = json.dumps({"tool": {"name": benchmark.tool_name, "revision": benchmark.tool_version}, \
+                                        "generator": "benchmark.vcloud.py"})
 
         # install cloud and dependencies
-        ant = subprocess.Popen(["ant", "resolve-benchmark-dependencies"], shell=Util.isWindows())
+        ant = subprocess.Popen(["ant", "resolve-benchmark-dependencies"],
+                               cwd=_ROOT_DIR,
+                               shell=util.is_windows())
         ant.communicate()
         ant.wait()
 
@@ -67,46 +84,53 @@ def executeBenchmarkInCloud(benchmark, outputHandler, justReprocessResults):
             logLevel =  "FINER"
         else:
             logLevel = "INFO"
-        heapSize = 100 + numberOfRuns//10 # 100 MB and 100 kB per run
-        libDir = os.path.abspath(os.path.join(os.path.curdir, "lib", "java-benchmark"))
-        cmdLine = ["java", "-Xmx"+str(heapSize)+"m", "-jar", os.path.join(libDir, "vcloud.jar"), "benchmark", "--loglevel", logLevel]
+        heapSize = benchmark.config.cloudClientHeap + numberOfRuns//10 # 100 MB and 100 kB per run
+        lib = os.path.join(_ROOT_DIR, "lib", "java-benchmark", "vcloud.jar")
+        cmdLine = ["java", "-Xmx"+str(heapSize)+"m", "-jar", lib, "benchmark", "--loglevel", logLevel, \
+                   "--run-collection-meta-information", meta_information, \
+                   "--environment", formatEnvironment(benchmark.environment()), \
+                   "--max-log-file-size", str(benchmark.config.maxLogfileSize), \
+                   "--debug", str(benchmark.config.debug)]
         if benchmark.config.cloudMaster:
             cmdLine.extend(["--master", benchmark.config.cloudMaster])
         if benchmark.config.debug:
             cmdLine.extend(["--print-new-files", "true"])
-            
-        wallTimeBefore = time.time()
-            
-        cloud = subprocess.Popen(cmdLine, stdin=subprocess.PIPE, shell=Util.isWindows())
+
+        walltime_before = time.time()
+
+        cloud = subprocess.Popen(cmdLine, stdin=subprocess.PIPE, shell=util.is_windows())
         try:
-            (out, err) = cloud.communicate(cloudInput.encode('utf-8'))
+            cloud.communicate(cloudInput.encode('utf-8'))
         except KeyboardInterrupt:
-            killScriptCloud()
+            stop()
         returnCode = cloud.wait()
 
-        wallTimeAfter = time.time()
-        usedWallTime = wallTimeAfter - wallTimeBefore
+        walltime_after = time.time()
+        usedWallTime = walltime_after - walltime_before
 
         if returnCode:
             if STOPPED_BY_INTERRUPT:
-                outputHandler.setError('interrupted')
+                output_handler.set_error('interrupted')
             else:
                 errorMsg = "Cloud return code: {0}".format(returnCode)
-                logging.warn(errorMsg)
-                outputHandler.setError(errorMsg)
+                logging.warning(errorMsg)
+                output_handler.set_error(errorMsg)
     else:
-        returnCode = 0    
+        returnCode = 0
+        usedWallTime = None
 
-    handleCloudResults(benchmark, outputHandler, usedWallTime)
+    handleCloudResults(benchmark, output_handler, usedWallTime)
 
     return returnCode
 
 
-def killScriptCloud():
+def stop():
     global STOPPED_BY_INTERRUPT
     STOPPED_BY_INTERRUPT = True
     # kill cloud-client, should be done automatically, when the subprocess is aborted
 
+def formatEnvironment(environment):
+    return ",".join(k + "=" + v for k,v in environment.items())
 
 def toTabList(l):
     return "\t".join(map(str, l))
@@ -118,13 +142,12 @@ def getCloudInput(benchmark):
     (workingDir, toolpaths) = getToolDataForCloud(benchmark)
 
     # prepare cloud input, we make all paths absolute, TODO necessary?
-    outputDir = benchmark.logFolder
+    outputDir = benchmark.log_folder
     absOutputDir = os.path.abspath(outputDir)
     absWorkingDir = os.path.abspath(workingDir)
-    absScriptsPath = os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir))
     absToolpaths = list(map(os.path.abspath, toolpaths))
     absSourceFiles = list(map(os.path.abspath, sourceFiles))
-    absBaseDir = Util.commonBaseDir(absSourceFiles + absToolpaths + [absScriptsPath])
+    absBaseDir = util.common_base_dir(absSourceFiles + absToolpaths)
 
     if absBaseDir == "": sys.exit("No common base dir found.")
 
@@ -135,12 +158,12 @@ def getCloudInput(benchmark):
     # build the input for the cloud,
     # see external vcloud/README.txt for details.
     cloudInput = [
-                toTabList(absToolpaths + [absScriptsPath]),
+                toTabList(absToolpaths),
                 toTabList([absBaseDir, absOutputDir, absWorkingDir]),
                 toTabList(requirements)
             ]
-    if benchmark.resultFilesPattern:
-        cloudInput.append(benchmark.resultFilesPattern)
+    if benchmark.result_files_pattern:
+        cloudInput.append(benchmark.result_files_pattern)
 
     cloudInput.extend([
                 toTabList(numOfRunDefLinesAndPriorityStr),
@@ -154,47 +177,40 @@ def getBenchmarkDataForCloud(benchmark):
 
     # get requirements
     r = benchmark.requirements
-    requirements = [DEFAULT_CLOUD_MEMORY_REQUIREMENT if r.memory is None else r.memory,
-                    DEFAULT_CLOUD_CPUCORE_REQUIREMENT if r.cpuCores is None else r.cpuCores,
-                    DEFAULT_CLOUD_CPUMODEL_REQUIREMENT if r.cpuModel is None else r.cpuModel]
+    requirements = [bytes_to_mb(DEFAULT_CLOUD_MEMORY_REQUIREMENT if r.memory is None else r.memory),
+                    DEFAULT_CLOUD_CPUCORE_REQUIREMENT if r.cpu_cores is None else r.cpu_cores,
+                    DEFAULT_CLOUD_CPUMODEL_REQUIREMENT if r.cpu_model is None else r.cpu_model]
 
     # get limits and number of Runs
     timeLimit = benchmark.rlimits.get(TIMELIMIT, DEFAULT_CLOUD_TIMELIMIT)
-    memLimit  = benchmark.rlimits.get(MEMLIMIT,  DEFAULT_CLOUD_MEMLIMIT)
+    memLimit  = bytes_to_mb(benchmark.rlimits.get(MEMLIMIT,  DEFAULT_CLOUD_MEMLIMIT))
     coreLimit = benchmark.rlimits.get(CORELIMIT, None)
-    numberOfRuns = sum(len(runSet.runs) for runSet in benchmark.runSets if runSet.shouldBeExecuted())
+    numberOfRuns = sum(len(runSet.runs) for runSet in benchmark.run_sets if runSet.should_be_executed())
     limitsAndNumRuns = [numberOfRuns, timeLimit, memLimit]
     if coreLimit is not None: limitsAndNumRuns.append(coreLimit)
-
-    # get tool-specific environment
-    env = benchmark.getEnvironments()
 
     # get Runs with args and sourcefiles
     sourceFiles = []
     runDefinitions = []
-    for runSet in benchmark.runSets:
-        if not runSet.shouldBeExecuted(): continue
+    for runSet in benchmark.run_sets:
+        if not runSet.should_be_executed(): continue
         if STOPPED_BY_INTERRUPT: break
 
         # get runs
         for run in runSet.runs:
-            cmdline = run.getCmdline()
-            cmdline = list(map(Util.forceLinuxPath, cmdline))
+            cmdline = run.cmdline()
+            cmdline = list(map(util.force_linux_path, cmdline))
 
             # we assume, that VCloud-client only splits its input at tabs,
             # so we can use all other chars for the info, that is needed to run the tool.
-            # we build a string-representation of all this info (it's a map),
-            # that can be parsed with python again in cloudRunexecutor.py (this is very easy with eval()) .
-            argMap = {"args":cmdline, "env":env,
-                      "debug":benchmark.config.debug, "maxLogfileSize":benchmark.config.maxLogfileSize}
-            argString = repr(argMap)
+            argString = json.dumps(cmdline) 
             assert not "\t" in argString # cannot call toTabList(), if there is a tab
 
-            logFile = os.path.relpath(run.logFile, benchmark.logFolder)
+            log_file = os.path.relpath(run.log_file, benchmark.log_folder)
             if os.path.exists(run.identifier):
-                runDefinitions.append(toTabList([argString, logFile] + run.sourcefiles + run.requiredFiles))
+                runDefinitions.append(toTabList([argString, log_file] + run.sourcefiles + run.required_files))
             else:
-                runDefinitions.append(toTabList([argString, logFile] + run.requiredFiles))
+                runDefinitions.append(toTabList([argString, log_file] + run.required_files))
             sourceFiles.extend(run.sourcefiles)
 
     if not runDefinitions: sys.exit("Benchmark has nothing to run.")
@@ -204,12 +220,12 @@ def getBenchmarkDataForCloud(benchmark):
 
 def getToolDataForCloud(benchmark):
 
-    workingDir = benchmark.workingDirectory()
+    workingDir = benchmark.working_directory()
     if not os.path.isdir(workingDir):
         sys.exit("Missing working directory {0}, cannot run tool.".format(workingDir))
     logging.debug("Working dir: " + workingDir)
 
-    toolpaths = benchmark.requiredFiles()
+    toolpaths = benchmark.required_files()
     for file in toolpaths:
         if not os.path.exists(file):
             sys.exit("Missing file {0}, cannot run benchmark within cloud.".format(os.path.normpath(file)))
@@ -217,71 +233,64 @@ def getToolDataForCloud(benchmark):
     return (workingDir, toolpaths)
 
 
-def handleCloudResults(benchmark, outputHandler, usedWallTime):
+def handleCloudResults(benchmark, output_handler, usedWallTime):
 
-    outputDir = benchmark.logFolder
+    outputDir = benchmark.log_folder
     if not os.path.isdir(outputDir) or not os.listdir(outputDir):
         # outputDir does not exist or is empty
-        logging.warning("Cloud produced no results. Output-directory is missing or empty: {0}".format(outputDir))
+        logging.warning("Cloud produced no results. Output-directory is missing or empty: %s", outputDir)
 
     # Write worker host informations in xml
-    filePath = os.path.join(outputDir, "hostInformation.txt")
-    parseAndSetCloudWorkerHostInformation(filePath, outputHandler)
+    parseAndSetCloudWorkerHostInformation(outputDir, output_handler, benchmark)
 
     # write results in runs and handle output after all runs are done
     executedAllRuns = True
     runsProducedErrorOutput = False
-    for runSet in benchmark.runSets:
-        if not runSet.shouldBeExecuted():
-            outputHandler.outputForSkippingRunSet(runSet)
+    for runSet in benchmark.run_sets:
+        if not runSet.should_be_executed():
+            output_handler.output_for_skipping_run_set(runSet)
             continue
 
-        outputHandler.outputBeforeRunSet(runSet)
+        output_handler.output_before_run_set(runSet)
 
         for run in runSet.runs:
-            dataFile = run.logFile + ".data"
-            if os.path.exists(dataFile):
+            dataFile = run.log_file + ".data"
+            if os.path.exists(dataFile) and os.path.exists(run.log_file):
                 try:
-                    (run.cpuTime, run.wallTime, returnValue, values) = parseCloudRunResultFile(dataFile)
-                    run.values.update(values)
-                    if returnValue is not None and not benchmark.config.debug:
-                        # Do not delete .data file if there was some problem
+                    values = parseCloudRunResultFile(dataFile)
+                    if not benchmark.config.debug:
                         os.remove(dataFile)
                 except IOError as e:
-                    logging.warning("Cannot extract measured values from output for file {0}: {1}".format(
-                                    outputHandler.formatSourceFileName(run.identifier), e))
-                    outputHandler.allCreatedFiles.append(dataFile)
+                    logging.warning("Cannot extract measured values from output for file %s: %s",
+                                    run.identifier, e)
+                    output_handler.all_created_files.add(dataFile)
                     executedAllRuns = False
-                    returnValue = None
+                else:
+                    output_handler.output_before_run(run)
+                    run.set_result(values, ['host'])
+                    output_handler.output_after_run(run)
             else:
-                logging.warning("No results exist for file {0}.".format(outputHandler.formatSourceFileName(run.identifier)))
+                logging.warning("No results exist for file %s.", run.identifier)
                 executedAllRuns = False
-                returnValue = None
 
-            if os.path.exists(run.logFile + ".stdError"):
+            if os.path.exists(run.log_file + ".stdError"):
                 runsProducedErrorOutput = True
 
-            outputHandler.outputBeforeRun(run)
+        output_handler.output_after_run_set(runSet, walltime=usedWallTime)
 
-            run.afterExecution(returnValue)
-            outputHandler.outputAfterRun(run)
-
-        outputHandler.outputAfterRunSet(runSet, wallTime=usedWallTime)
-
-    outputHandler.outputAfterBenchmark(STOPPED_BY_INTERRUPT)
+    output_handler.output_after_benchmark(STOPPED_BY_INTERRUPT)
 
     if not executedAllRuns:
         logging.warning("Some expected result files could not be found!")
     if runsProducedErrorOutput and not benchmark.config.debug:
-        logging.warning("Some runs produced unexpected warnings on stderr, please check the {0} files!"
-                        .format(os.path.join(outputDir, '*.stdError')))
+        logging.warning("Some runs produced unexpected warnings on stderr, please check the %s files!",
+                        os.path.join(outputDir, '*.stdError'))
 
 
-def parseAndSetCloudWorkerHostInformation(filePath, outputHandler):
+def parseAndSetCloudWorkerHostInformation(outputDir, output_handler, benchmark):
+    filePath = os.path.join(outputDir, "hostInformation.txt")
     try:
         with open(filePath, 'rt') as file:
-            outputHandler.allCreatedFiles.append(filePath)
-
             # Parse first part of information about hosts until first blank line
             while True:
                 line = file.readline().strip()
@@ -293,20 +302,24 @@ def parseAndSetCloudWorkerHostInformation(filePath, outputHandler):
                 cpuName = file.readline().split("=")[-1].strip()
                 frequency = file.readline().split("=")[-1].strip()
                 cores = file.readline().split("=")[-1].strip()
-                outputHandler.storeSystemInfo(osName, cpuName, cores, frequency, memory, name)
+                output_handler.store_system_info(osName, cpuName, cores, frequency, memory, name)
 
             # Ignore second part of information about runs
             # (we read the run-to-host mapping from the .data file for each run).
 
+        if benchmark.config.debug:
+            output_handler.all_created_files.add(filePath)
+        else:
+            os.remove(filePath)
     except IOError:
         logging.warning("Host information file not found: " + filePath)
 
 
+IGNORED_VALUES = set(['command', 'timeLimit', 'coreLimit', 'returnvalue', 'exitsignal'])
+"""result values that are ignored because they are redundant"""
+
 def parseCloudRunResultFile(filePath):
     values = {}
-    cpuTime = None
-    wallTime = None
-    returnValue = None
 
     def parseTimeValue(s):
         if s[-1] != 's':
@@ -315,25 +328,24 @@ def parseCloudRunResultFile(filePath):
 
     with open(filePath, 'rt') as file:
         for line in file:
-            (key, value) = line.split("=", 2)
+            (key, value) = line.split("=", 1)
             value = value.strip()
             if key == 'cputime':
-                cpuTime = parseTimeValue(value)
+                values['cputime'] = parseTimeValue(value)
             elif key == 'walltime':
-                wallTime = parseTimeValue(value)
-            elif key == 'cpuTime':
-                cpuTime = float(value)
-            elif key == 'wallTime':
-                wallTime = float(value)
+                values['walltime'] = parseTimeValue(value)
             elif key == 'memory':
-                values['memUsage'] = value
+                values['memory'] = int(value)
             elif key == 'exitcode':
-                returnValue = int(value)
-                values['@exitcode'] = value
-            elif key == "host":
+                values['exitcode'] = int(value)
+            elif (key == "host" or key == "terminationreason" or
+                  key.startswith("energy-") or key.startswith("cputime-cpu")):
                 values[key] = value
-            else:
-                # "@" means value is hidden normally
-                values["@vcloud-" + key] = value
+            elif key not in IGNORED_VALUES:
+                values["vcloud-" + key] = value
+    return values
 
-    return (cpuTime, wallTime, returnValue, values)
+def bytes_to_mb(mb):
+    if mb is None:
+        return None
+    return int(mb / 1000 / 1000)
